rename CatSimInterface to CatSimDatabaseEmulator
